requirements.txt
Django==1.7.7
django-bootstrap3-datetimepicker==2.2.3
django-forms-bootstrap==3.0.1
django-phonenumber-field==0.7.2
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.0.0
static3==0.5.1
phonenumbers==7.0.2
psycopg2==2.6
wsgiref==0.1.2
geopy==1.9.1
django-twilio==0.4
twilio==3.6.3
requests==2.6.2
lxml==3.3.3
cssselect==0.9.1